requirements.txt
opencv-python==4.8.0.74
colorlog
numpy
Pillow
ipython
imageio
tqdm
scipy
lark
openai
matplotlib
scikit-image
wandb
rtpt
scikit-learn
grad-cam
einops
stable_baselines3
gym
numpy
debugpy
faiss-cpu
